The Application supports one instance of a face recognizer. The user has to choose and apply a face recognizer in the "File/Select face recognizer" menu.
The user can train the applied face recognizer with images. The
application finds all of the files from a selected directory, which can be described with the regular
expression: \w+_\w+_\d+.jpg. After the training the face recognizer is
saved to a file which is selected by the user. Also a name label file is
generated.

During the database creation process the preprocessed images are saved to
a folder named: preprocessed

For testing purposes I used the images downloaded from
vis-www.cs.umass.edu/lfw/
I used the names beginning with A and B. With the EigenFaceRecognizer
num_components set to 80, the Application uses approximately 9 GB of
RAM.